+++ SeaGame/src/TransformComponent.cpp
@@ -23,9 +23,9 @@ void TransformComponent::setRotation(float newRot)
 	this->rotation = newRot;
 	// Ensure the rotation is always between 0 and 2pi
 	while (this->rotation < 0)
-		this->rotation += 2 * M_PI;
+		this->rotation += (float)(2 * M_PI);
 	while (this->rotation > 2 * M_PI)
-		this->rotation -= 2 * M_PI;
+		this->rotation -= (float)(2 * M_PI);
 }
 
 std::pair<sf::Vector2f, float> TransformComponent::getDifference(TransformComponent other)
@@ -35,14 +35,14 @@ std::pair<sf::Vector2f, float> TransformComponent::getDifference(TransformCompon
 	// So the next 3 lines convert it to [0, 2pi]
 	float angle = atan2(diff.y, diff.x);
 	if (angle < 0)
-		angle += 2 * M_PI;
+		angle += (float)(2 * M_PI);
 	// Get difference between this entity's rotation and the angle to the other
 	angle -= this->getRotation();
 	// If the angle is greater than pi, it is shorter to go the negative direction
 	if (angle > M_PI)
-		angle -= 2 * M_PI;
+		angle -= (float)(2 * M_PI);
 	else if (angle < -M_PI)
-		angle += 2 * M_PI;
+		angle += (float)(2 * M_PI);
 	return { diff, angle };
 
 }
